financial_insights/requirements.txt
chromadb==0.4.24
fpdf2==2.7.9
fuzzywuzzy==0.18.0
kaleido==0.2.1
langchain==0.2.13
langchain-community==0.2.6
langchain-experimental==0.0.60
llama-index-readers-sec-filings==0.1.5
mypy==1.11.0
pandasai==2.2.10
plotly == 5.22.0
pydantic==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
secedgar==0.5.0
sentence-transformers==3.0.1
sec-downloader==0.11.1
sseclient-py==1.8.0
streamlit==1.36.0
streamlit_extras==0.4.3
types-PyYAML==6.0.12.20240311
yfinance==0.2.41